refactor(core): split kernel event modules
Split the kernel event catalog, EventBus implementation, and scoped bus into separate modules while preserving the existing events.ts barrel. Add a kernel boundary test to keep high-level runtime imports out of the kernel.
